+++ lib/src/network/util_https.dart
@@ -18,13 +18,17 @@ class UtilHttps {
   /// encoded according to the enum specification.
   /// * [jwt] : The jwt. It is inserted into the Authorization header.
   /// * [timeout] : The response timeout.
+  /// * [adjustTiming] : Specify true to automatically adjust the timing.
+  /// * [intervalMSec] : The minimum interval between calls that is
+  /// automatically adjusted if adjustTiming is True.
+  /// If consecutive calls are made earlier than this,
+  /// they will wait until this interval before being executed.
   static Future<ServerResponse> post(
-    String url,
-    Map<String, dynamic> body,
-    EnumPostEncodeType type, {
-    String? jwt,
-    Duration timeout = const Duration(seconds: 10),
-  }) async {
+      String url, Map<String, dynamic> body, EnumPostEncodeType type,
+      {String? jwt,
+      Duration timeout = const Duration(seconds: 10),
+      bool adjustTiming = true,
+      intervalMSec = 1200}) async {
     Map<String, String> headers = {};
     if (jwt != null) {
       headers['Authorization'] = 'Bearer $jwt';
@@ -33,10 +37,15 @@ class UtilHttps {
       case EnumPostEncodeType.urlEncoded:
         headers['Content-Type'] = 'application/x-www-form-urlencoded';
         return customPost(url, Uri(queryParameters: body).query, headers,
-            timeout: timeout);
+            timeout: timeout,
+            adjustTiming: adjustTiming,
+            intervalMSec: intervalMSec);
       case EnumPostEncodeType.json:
         headers['Content-Type'] = 'application/json';
-        return customPost(url, jsonEncode(body), headers, timeout: timeout);
+        return customPost(url, jsonEncode(body), headers,
+            timeout: timeout,
+            adjustTiming: adjustTiming,
+            intervalMSec: intervalMSec);
     }
   }
 
@@ -53,14 +62,21 @@ class UtilHttps {
   /// * [headers] : HTTP headers.
   /// * [encoding] : The data encoding.
   /// * [timeout] : The response timeout.
+  /// * [adjustTiming] : Specify true to automatically adjust the timing.
+  /// * [intervalMSec] : The minimum interval between calls that is
+  /// automatically adjusted if adjustTiming is True.
+  /// If consecutive calls are made earlier than this,
+  /// they will wait until this interval before being executed.
   static Future<ServerResponse> customPost(
-    String url,
-    Object? body,
-    Map<String, String> headers, {
-    Encoding? encoding,
-    Duration timeout = const Duration(seconds: 10),
-  }) async {
+      String url, Object? body, Map<String, String> headers,
+      {Encoding? encoding,
+      Duration timeout = const Duration(seconds: 10),
+      bool adjustTiming = true,
+      intervalMSec = 1200}) async {
     try {
+      if (adjustTiming) {
+        await TimingManager().adjustTiming(intervalMs: intervalMSec);
+      }
       final String httpsURL = UtilCheckURL.validateHttpsUrl(url);
       final http.Response r = await http
           .post(Uri.parse(httpsURL),