+++ lib/src/ropc_client_for_native.dart
@@ -22,7 +22,7 @@ import 'package:simple_jwt_manager/src/static_fields/f_json_keys_to_server.dart'
 class ROPCClientForNative {
   // static parameters
   static const String className = "ROPCClientForNative";
-  static const int version = 1;
+  static const int version = 2;
 
   // parameters
   late final String _registerUrl;
@@ -34,6 +34,7 @@ class ROPCClientForNative {
       _badCertificateCallback;
   late final Duration _connectionTimeout;
   late final Duration _responseTimeout;
+  late final int refreshMarginMs;
 
   // tokens
   String? _accessToken;
@@ -60,6 +61,10 @@ class ROPCClientForNative {
   /// that uses a self-signed certificate.
   /// * [savedData] : If there is token information previously saved by
   /// this class's toDict function, you can restore the token by setting it.
+  /// * [refreshMarginMs] : The access token expiration time will be estimated
+  /// early by the margin you set here.
+  /// By default, if the access token is due to expire within 30 seconds,
+  /// it will be automatically refreshed using a refresh token.
   ROPCClientForNative(
       {required String registerURL,
       required String signInURL,
@@ -70,7 +75,8 @@ class ROPCClientForNative {
       Duration? responseTimeout,
       bool Function(X509Certificate cert, String host, int port)?
           badCertificateCallback,
-      Map<String, dynamic>? savedData}) {
+      Map<String, dynamic>? savedData,
+      this.refreshMarginMs = 30 * 1000}) {
     _registerUrl = UtilCheckURL.validateHttpsUrl(registerURL);
     _signInUrl = UtilCheckURL.validateHttpsUrl(signInURL);
     _refreshUrl = UtilCheckURL.validateHttpsUrl(refreshURL);
@@ -200,7 +206,8 @@ class ROPCClientForNative {
         EnumPostEncodeType.json,
         badCertificateCallback: _badCertificateCallback,
         connectionTimeout: _connectionTimeout,
-        responseTimeout: _responseTimeout);
+        responseTimeout: _responseTimeout,
+        adjustTiming: false);
     switch (r.resultStatus) {
       case EnumSeverResponseStatus.success:
         try {
@@ -246,7 +253,8 @@ class ROPCClientForNative {
         EnumPostEncodeType.json,
         badCertificateCallback: _badCertificateCallback,
         connectionTimeout: _connectionTimeout,
-        responseTimeout: _responseTimeout);
+        responseTimeout: _responseTimeout,
+        adjustTiming: false);
     switch (r.resultStatus) {
       case EnumSeverResponseStatus.success:
         _clearToken();
@@ -306,7 +314,8 @@ class ROPCClientForNative {
         EnumPostEncodeType.urlEncoded,
         badCertificateCallback: _badCertificateCallback,
         connectionTimeout: _connectionTimeout,
-        responseTimeout: _responseTimeout);
+        responseTimeout: _responseTimeout,
+        adjustTiming: false);
     switch (r.resultStatus) {
       case EnumSeverResponseStatus.success:
         // トークンを取得して保存
@@ -332,11 +341,14 @@ class ROPCClientForNative {
   /// If an error occurs along the way,
   /// an error will simply be returned,
   /// which may make it difficult to understand the situation.
+  /// To find out which token failed to be revoked when an error occurred,
+  /// check the tokens held by the client.
   ///
   /// (ja) RFC 7009の仕様の通り、サインアウト処理を行います。
   /// これは簡易版で、リフレッシュトークンとアクセストークンを連続で失効処理します。
   /// 途中で何らかのエラーが発生した場合は、単にエラーが返されます。
-  /// 処理が失敗した場合、どちらの失効リクエストが失敗したのかを判断するには、
+  /// エラー時にどのトークンの失効に失敗したのかを調べるには、
+  /// クライアントの保持するトークンを確認してください。
   Future<ServerResponse> signOutAllTokens() async {
     ServerResponse res = await signOut(isRefreshToken: true);
     switch (res.resultStatus) {
@@ -389,7 +401,8 @@ class ROPCClientForNative {
         _revokeURL, target, EnumPostEncodeType.urlEncoded,
         badCertificateCallback: _badCertificateCallback,
         connectionTimeout: _connectionTimeout,
-        responseTimeout: _responseTimeout);
+        responseTimeout: _responseTimeout,
+        adjustTiming: false);
     switch (r.resultStatus) {
       case EnumSeverResponseStatus.success:
         if (isRefreshToken) {
@@ -438,7 +451,8 @@ class ROPCClientForNative {
   /// トークンの期限切れをチェックします。
   bool _isTokenExpired() {
     if (_accessTokenExpireUnixMS == null) return true;
-    return DateTime.now().millisecondsSinceEpoch > _accessTokenExpireUnixMS!;
+    return DateTime.now().millisecondsSinceEpoch >
+        _accessTokenExpireUnixMS! - refreshMarginMs;
   }
 
   /// リフレッシュトークンを使用して新しいトークンを取得し、キャッシュします。
@@ -455,7 +469,8 @@ class ROPCClientForNative {
         EnumPostEncodeType.urlEncoded,
         badCertificateCallback: _badCertificateCallback,
         connectionTimeout: _connectionTimeout,
-        responseTimeout: _responseTimeout);
+        responseTimeout: _responseTimeout,
+        adjustTiming: false);
     switch (r.resultStatus) {
       case EnumSeverResponseStatus.success:
         try {